refactor: split plan executor components
